fixing header info

+++ src/forthwith.h
@@ -80,7 +80,7 @@ typedef struct forthwith_str_stack { /**< FORTH environment */
 } fw_ctx_str_stack_t;
 
 struct user_ptr {
-  uint16_t *data;
+  uint8_t *data;
   uint16_t elem_count;
   uint16_t elem_size;
   uint16_t elem_idx;
